Exercise3 _ 개선코드

+++ Exercise/src/Ex_ch03/Exercise3_08.java
@@ -15,7 +15,8 @@ public class Exercise3_08 {
 		long l = 3000L * 3000 * 3000;
 		float f2 = 0.1f;
 		double d = 0.1f;
-		boolean result = d == f2;
+		 // d->f 명시적 형변환 해줘야 함
+		boolean result = (float)d == f2;
 		System.out.println("c=" + c);
 		System.out.println("ch=" + ch);
 		System.out.println("f=" + f);
